Core complete

+++ src/main/java/com/booleanuk/api/game/GameRepository.java
@@ -2,6 +2,6 @@ package com.booleanuk.api.game;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GameRepository extends JpaRepository<Game, Long> {
+public interface GameRepository extends JpaRepository<Game, Integer> {
 
 }
